feat: a bit more adc calibration
