+++ firmware/development/kampela_experiments_efm32pg23/app3/src/devices/mod.rs
@@ -1,0 +1,1 @@
+pub mod display;
