+++ firmware/development/kampela_experiments_efm32pg23/app3/src/lib.rs
@@ -12,8 +12,8 @@
 extern crate alloc;
 
 pub mod init;
-pub mod gpio_pins;
-
+mod peripherals;
+mod devices;
 pub mod draw;
 pub mod se;
 use kampela_ui as ui;
