+++ firmware/development/kampela_experiments_efm32pg23/app3/src/peripherals/adc.rs
@@ -2,6 +2,11 @@
 
 use efm32pg23_fix::Peripherals;
 
+//debug tools
+//TODO: remove!
+use crate::draw::burning_tank;
+use alloc::format;
+
 /// Initialize ADC
 ///
 /// assumes that CMU clock is enabled and does not check it
@@ -56,25 +61,65 @@ pub fn init_adc(peripherals: &mut Peripherals) {
         });
 
     // calibration data
-    let ui_gain_cana = peripherals
+    //
+    // things are complicated...
+    let ui_gain = peripherals
         .DEVINFO
         .iadc0gain0
         .read()
         .gaincana1()
         .bits();
-
-    peripherals
-        .IADC0_S
-        .scale0
-        .write(|w_reg| w_reg.gain13lsb().variant(ui_gain_cana));
-
-    let offset = peripherals
+    let offset1 = peripherals
         .DEVINFO
         .iadc0normaloffsetcal1
         .read()
         .offsetana3norm()
-        .bits();
+        .bits() as i16; // C reference did this
+    let offset0 = peripherals
+        .DEVINFO
+        .iadc0normaloffsetcal0
+        .read()
+        .bits() as i16; // C reference did this
 
+    // do some floating point
+    let offset = offset1 / 16 + offset0;
+    let offset_f = offset as f32 + 87505.45;
+    let offset_f = (ui_gain as f32 / 5416.198) * offset_f - 524288.0;
+    let offset = offset_f as i32;
+
+    // mess with bytes now
+    let ui_gain_sign = ui_gain & 0x8000;
+    let ui_gain_value = ui_gain & 0x1FFF;
+    let offset_truncated = match offset {
+        i32::MIN..=-0x20001 => -0x20000,
+        -0x20000..=0x1FFFF => offset,
+        0x20000..=i32::MAX => 0x1FFFF,
+    };
+
+    peripherals
+        .IADC0_S
+        .scale0
+        .write(|w_reg| {
+            let prefab = w_reg
+            .offset().variant(offset_truncated as u32)
+            .gain13lsb().variant(ui_gain)
+            .gain3msb();
+            if ui_gain_sign == 0x8000 {
+                prefab.gain100()
+            } else {
+                prefab.gain011()
+            }
+        });
+
+    peripherals
+        .IADC0_S
+        .sched0
+        .write(|w_reg| w_reg.prescale().variant(1));
+
+    /*
+    burning_tank(peripherals, format!("gain: {}; offset1: {}, offest0: {}, offset: {}", ui_gain, offset1, offset0, offset));
+    loop{}
+    */
 }
 
 /// Disable ADC
