+++ firmware/development/kampela_experiments_efm32pg23/app3/src/peripherals/mod.rs
@@ -1,0 +1,4 @@
+pub mod adc;
+pub mod cmu;
+pub mod gpio_pins;
+pub mod usart;
